+++ packages/sources/tradingeconomics/src/config.ts
@@ -1,31 +1,19 @@
-import { Requester } from '@chainlink/ea-bootstrap'
-import { Config as BaseConfig } from '@chainlink/types'
 import { util } from '@chainlink/ea-bootstrap'
 
-export const DEFAULT_ENDPOINT = 'markets'
-export const DEFAULT_BASE_URL = 'https://api.tradingeconomics.com'
-
-const DEFAULT_WS_TIMEOUT = 5000
-
-export type Config = BaseConfig & {
-  wsTimeout: number
-  apiClientKey: string
+export type Config = {
+  url: string
+  key: string
+  secret: string
+  symbols: string
+  reconnectTimeout: number
 }
 
 export const makeConfig = (prefix?: string): Config => {
-  const config = Requester.getDefaultConfig(prefix) as Config
-  config.api.baseURL = config.api.baseURL || DEFAULT_BASE_URL
-
-  config.wsTimeout = Number(util.getEnv('WS_TIMEOUT')) || DEFAULT_WS_TIMEOUT
-
-  const CLIENT_KEY = util.getRequiredEnv('API_CLIENT_KEY')
-  config.apiClientKey = CLIENT_KEY
-  const CLIENT_SECRET = util.getRequiredEnv('API_CLIENT_SECRET')
-  config.apiKey = CLIENT_SECRET
-
-  config.api.params = {
-    c: `${CLIENT_KEY}:${CLIENT_SECRET}`,
+  return {
+    url: util.getEnv('API_URL', prefix) || 'ws://stream.tradingeconomics.com/',
+    key: util.getRequiredEnv('API_CLIENT_KEY', prefix),
+    secret: util.getRequiredEnv('API_CLIENT_SECRET', prefix),
+    symbols: util.getRequiredEnv('SYMBOLS', prefix),
+    reconnectTimeout: Number(util.getEnv('RECONNECT_TIMEOUT', prefix) || 3000),
   }
-
-  return config
 }