+++ packages/sources/tradingeconomics/test/adapter.test.ts
@@ -1,56 +1,19 @@
-import { assert } from 'chai'
 import { Requester } from '@chainlink/ea-bootstrap'
-import { assertSuccess, assertError } from '@chainlink/ea-test-helpers'
+import { assertError } from '@chainlink/ea-test-helpers'
 import { AdapterRequest } from '@chainlink/types'
 import { makeExecute } from '../src/adapter'
 
 describe('execute', () => {
   const jobID = '1'
-  const execute = makeExecute()
-  if (!process.env.API_CLIENT_KEY) process.env.API_CLIENT_KEY = 'test_client_key'
-  if (!process.env.API_CLIENT_SECRET) process.env.API_CLIENT_SECRET = 'test_client_secret'
-
-  context('successful calls @integration', () => {
-    const requests = [
-      { name: 'id not supplied', testData: { data: { base: 'FTSE' } } },
-      { name: 'base', testData: { id: jobID, data: { base: 'FTSE' } } },
-      { name: 'from/to', testData: { id: jobID, data: { from: 'FTSE' } } },
-    ]
-
-    requests.forEach((req) => {
-      it(`${req.name}`, async () => {
-        const data = await execute(req.testData as AdapterRequest)
-        assertSuccess({ expected: 200, actual: data.statusCode }, data, jobID)
-        assert.isAbove(data.result, 0)
-        assert.isAbove(data.data.result, 0)
-      })
-    })
-  })
+  const execute = makeExecute({ key: '', reconnectTimeout: 0, secret: '', symbols: '', url: '' })
 
   context('validation error', () => {
     const requests = [
       { name: 'empty body', testData: {} },
       { name: 'empty data', testData: { data: {} } },
-      { name: 'base not supplied', testData: { id: jobID, data: {} } },
-    ]
-
-    requests.forEach((req) => {
-      it(`${req.name}`, async () => {
-        try {
-          await execute(req.testData as AdapterRequest)
-        } catch (error) {
-          const errorResp = Requester.errored(jobID, error)
-          assertError({ expected: 400, actual: errorResp.statusCode }, errorResp, jobID)
-        }
-      })
-    })
-  })
-
-  context('error calls @integration', () => {
-    const requests = [
       {
-        name: 'unknown base',
-        testData: { id: jobID, data: { base: 'not_real' } },
+        name: 'base not supplied',
+        testData: { id: jobID, data: { quote: 'USD' } },
       },
     ]
 
@@ -60,7 +23,7 @@ describe('execute', () => {
           await execute(req.testData as AdapterRequest)
         } catch (error) {
           const errorResp = Requester.errored(jobID, error)
-          assertError({ expected: 500, actual: errorResp.statusCode }, errorResp, jobID)
+          assertError({ expected: 400, actual: errorResp.statusCode }, errorResp, jobID)
         }
       })
     })